Update cookbook
This commit is to update the cookbook to the latest at the moment,
it updated the Berksfile and validate that the right cookbook name
for `golang_test` is well configured.

Changed the provider with a nice output. Though it need more work.

Also deleted some stuff from `.kitchen.yml` and added chef-zero
instead of chef-solo.