Fix source installation issues
Avoid conflicts of installing the source build into the same directory as the
bootstrap go. Instead, install the bootstrap go in a known location that will
not cause conflicts and also clean it up after a successful build.

This also adds a install-dir test suite to ensure that works as well.

Signed-off-by: Lance Albertson <lance@osuosl.org>